Module Dkml_install_api

module type Component_config = sig ... end

Each component must have a configuration module defined with the module type Component_config.

module Component_registry : sig ... end

The Component_registry is a global registry of all components that have been registered until now.