DKML Install API

Please visit the central documentation site for comprehensive documentation.

What follows is the OCaml library documentation when you pull in the Opam module dkml-install using something similar to the following in your *.opam file:

opam-version: "2.0"
depends: [
  "dkml-install" {>= "0.2.0"}
  (* ... *)
]

Library dkml-install.register

In your Dune file dune you can include the library as follows:

 (library
  (public_name your-component)
  (name your_component)
  (libraries dkml-install.register))

The entry point of this library is the module: Dkml_install_register.

Library dkml-install.api

Typically you get dkml-install.api automatically when you pull in dkml-install.register. If you need the API library by itself, in your Dune file dune you can include the library as follows:

 (library
  (libraries dkml-install.api))

The entry point of this library is the module: Dkml_install_api.